Moved LottieAnimation to its own module
