+++ sample-compose/build.gradle
@@ -50,7 +50,7 @@ androidExtensions {
 }
 
 dependencies {
-  implementation project(':lottie')
+  implementation project(':lottie-compose')
   implementation 'androidx.core:core-ktx:1.3.2'
   implementation 'androidx.multidex:multidex:2.0.1'
   implementation 'androidx.appcompat:appcompat:1.2.0'
