+++ sample-compose/src/main/java/com/airbnb/lottie/sample/compose/composables/Loader.kt
@@ -5,6 +5,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import com.airbnb.lottie.compose.LottieAnimation
+import com.airbnb.lottie.compose.LottieAnimationSpec
 import com.airbnb.lottie.sample.compose.R
 
 @Composable
